refactor: split document and module API routers
